refactoring the proxy interface for a cleaner version
