+++ .gitignore
@@ -148,3 +148,5 @@ test_set
 training_set_*
 leap_net/train/tf_logs/
 leap_net/train/model_saved/
+leap_net/proxy/model_saved/
+leap_net/proxy/tf_logs/
